nyt se ei ainakaan poista kaikkia tilauksia

+++ public/admin-profile/order-data.js
@@ -3,6 +3,8 @@
 const fetchOrders = async (userId) => {
     const orderCont = document.getElementById('wip-orders');
     const modalOrderCont = document.getElementById('edit-orders');
+    const dialog = document.getElementById('edit-orders-dialog')
+    const ordersHTML = [];
 
     const response = await fetch (`/order-data/wip/auth/${userId}`, {
         method: 'GET',
@@ -13,7 +15,9 @@ const fetchOrders = async (userId) => {
 
     console.log('fetchOrders reached');
     const orderHistory = await response.json();
-    console.log('fetchOrders', orderHistory);  
+    console.log('fetchOrders', orderHistory); 
+    
+    let orderToDeleteId = null;
 
     for (let order of orderHistory.orders) {
         // content box on the left
@@ -43,7 +47,7 @@ const fetchOrders = async (userId) => {
         orderDiv.appendChild(price);
     
         thisOrder.appendChild(orderDiv);
-        orderCont.appendChild(thisOrder);
+        ordersHTML.push(thisOrder);
 
         // orders modal
         var duplicateOrder = orderDiv.cloneNode(true);
@@ -85,6 +89,10 @@ const fetchOrders = async (userId) => {
 
         // buttons
         const removeBtn = document.getElementById('remove-icon');
+        const deletePopup = document.getElementById('delete-order');
+        const cancelBtn = deletePopup.querySelector('#cancel-order-delete');
+        const confirmBtn = deletePopup.querySelector('#confirm-order-delete');
+        
 
         doneIcon.addEventListener('click', async (e) => {
             e.preventDefault();
@@ -96,8 +104,6 @@ const fetchOrders = async (userId) => {
                 body: JSON.stringify({order_id: order.order_id, order_status: 'completed'})
             });
 
-            console.log(response, typeof response);
-
             if (response.status === 200) {
                 alert('Order marked as completed');
                 thisOrder.remove();
@@ -106,28 +112,59 @@ const fetchOrders = async (userId) => {
             } else {
                 alert("Something went wrong! Couldn't update order data");
             };
-        })
+        });
 
-        removeBtn.addEventListener('click', async (e) => {
+        removeBtn.addEventListener('click', (currentOrderId) => {
+            orderToDeleteId = order.order_id;
+
+            document.getElementById('order-id').innerText = orderToDeleteId;
+
+            deletePopup.classList.remove('hidden');
+            deletePopup.classList.add('open-popup');
+            dialog.classList.add('blur-background');
+        });
+
+        cancelBtn.addEventListener('click', () => {
+            orderToDeleteId = null;
+            deletePopup.classList.remove('open-popup');
+            deletePopup.classList.add('hidden');
+            dialog.classList.remove('blur-background');
+        });
+
+        confirmBtn.addEventListener('click', async (e) => {
             e.preventDefault();
-            const response = await fetch (`/order-data/wip/auth/${order.order_id}`, {
-                method: 'DELETE',
-                headers: {
-                    'Content-Type': 'application/json',
-                },
-            });
 
+            if (orderToDeleteId != null) {
+                const response = await fetch (`/order-data/wip/auth/${orderToDeleteId}`, {
+                    method: 'DELETE',
+                    headers: {
+                        'Content-Type': 'application/json',
+                    },
+                });
+    
+    
+                if (response.status === 204) {
+                    alert('Order deleted successfully');
+                    thisOrder.remove();
+                    modalOrder.remove();
+                    window.location.reload();
+                } else {
+                    alert("Something went wrong! Couldn't delete order data");
+                };
 
-            if (response.status === 204) {
-                alert('Order deleted successfully');
-                thisOrder.remove();
-                modalOrder.remove();
-                window.location.reload();
+                orderToDeleteId = null;
+                deletePopup.classList.remove('open-popup');
+                deletePopup.classList.add('hidden');
+                dialog.classList.remove('blur-background');
             } else {
-                alert("Something went wrong! Couldn't delete order data");
-            };
+                alert('No order selected for deletion!');
+            }
         })
-    }  
+    };
+    
+    const firstOrder = ordersHTML[0].cloneNode(true);
+    orderCont.appendChild(firstOrder);
+
 };
 
 export default fetchOrders;
